# swagger.yaml
openapi: 3.0.3
info:
  title: API RESTful with Node.js - Express - PostgreSQL - OpenAPI 3.0.3
  description: |-
    <h4> This is an API for managing users with Node.js, Express, and PostgreSQL. It provides endpoints to create, retrieve, update, and delete user records. The API follows the OpenAPI 3.0.3 specification.<h4>
  version: 1.0.0
servers:
  - url: "http://localhost:3010/"
    description: Servidor local
...
paths:
...
  /rolePermission:
    get:
      summary: Obtener todos los permisos de roles
      description: Obtiene todos los permisos de roles.
      operationId: getRolePermission
      tags:
        - Relación roles-permisos
      responses:
        "200":
          description: Permisos de roles obtenidos correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id_role_permission:
                      type: integer
                      example: 1
                    role:
                      type: object
                      properties:
                        id_role:
                          type: integer
                          example: 8
                        role:
                          type: string
                          example: "Administrador"
                    permission:
                      type: object
                      properties:
                        id_permission:
                          type: integer
                          example: 4
                        permission:
                          type: string
                          example: "usuario:Escritura"
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example:
                  {
                    error: "Error en el servidor al obtener los permisos de roles",
                  }

    post:
      summary: Crear un nuevo permiso de rol
      description: Crea un nuevo permiso de rol.
      operationId: createRolePermission
      tags:
        - Relación roles-permisos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_role:
                  type: integer
                  example: 1
                id_permission:
                  type: array
                  items:
                    type: integer
                  example: [1]
      responses:
        "201":
          description: Permiso de rol creado correctamente
          content:
            text/plain:
              schema:
                type: string
                example: { result: "Permiso de rol creado correctamente" }
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example:
                  { error: "Error en el servidor al crear el permiso de rol" }
  /rolePermission/{id}:
    get:
      summary: Obtener un permiso de rol por ID
      description: Obtiene un permiso de rol por su ID.
      operationId: getRolePermissionById
      tags:
        - Relación roles-permisos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Permiso de rol obtenido correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id_role_permission:
                    type: integer
                    example: 5
                  role:
                    type: object
                    properties:
                      id_role:
                        type: integer
                        example: 8
                      role:
                        type: string
                        example: "Administrador"
                  permission:
                    type: object
                    properties:
                      id_permission:
                        type: integer
                        example: 6
                      permission:
                        type: string
                        example: "usuario:Modificación"
                  created_at:
                    type: string
                    format: date-time
                    example: "2024-03-14T04:04:03.150Z"
                  updated_at:
                    type: string
                    format: date-time
                    example: "2024-03-14T04:04:03.150Z"
        "400":
          description: Error en la solicitud
          content:
            text/plain:
              schema:
                type: string
                example:
                  {
                    message: "El permiso de rol con el ID proporcionado no existe",
                  }
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example:
                  { error: "Error en el servidor al obtener el permiso de rol" }

    put:
      summary: Actualizar un permiso de rol por ID
      description: Actualiza un permiso de rol por su ID.
      operationId: updateRolePermission
      tags:
        - Relación roles-permisos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_role:
                  type: integer
                  example: 1
                id_permission:
                  type: integer
                  example: 1
      responses:
        "200":
          description: Permiso de rol actualizado correctamente
          content:
            text/plain:
              schema:
                type: string
                example: { result: "Permiso de rol actualizado correctamente" }
        "400":
          description: Error en la solicitud
          content:
            text/plain:
              schema:
                type: string
                example:
                  {
                    message: "El permiso de rol con el ID proporcionado no existe",
                  }
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example:
                  {
                    error: "Error en el servidor al actualizar el permiso de rol",
                  }

    delete:
      summary: Eliminar un permiso de rol por ID
      description: Elimina un permiso de rol por su ID.
      operationId: deleteRolePermission
      tags:
        - Relación roles-permisos
      parameters:
        - name: id
          in: path
          description: ID del permiso de rol a eliminar
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Permiso de rol eliminado correctamente
          content:
            text/plain:
              schema:
                type: string
                example: { result: "Permiso de rol eliminado correctamente" }
        "400":
          description: Error en la solicitud
          content:
            text/plain:
              schema:
                type: string
                example:
                  {
                    message: "El permiso de rol con el ID proporcionado no existe",
                  }
        "500":
          description: Error interno del servidor
          content:
            text/plain:
              schema:
                type: string
                example:
                  {
                    error: "Error en el servidor al eliminar el permiso de rol",
                  }
...